refactor: s/models/authentication/ akvo.rest.models
`from rest_framework.authentication import TokenAuthentication, BaseAuthentication`

So it makes sense to have the same module name too

Related to #4872
