+++ akvo/rest/views/program_results_geo.py
@@ -24,7 +24,7 @@ from rest_framework.response import Response
 from rest_framework.status import HTTP_403_FORBIDDEN
 
 from akvo.codelists.models import ResultType
-from akvo.rest.models import TastyTokenAuthentication
+from akvo.rest.authentication import TastyTokenAuthentication
 from akvo.rsr.models import Project, IndicatorPeriod, IndicatorPeriodData
 from akvo.rsr.models.result.utils import QUANTITATIVE, QUALITATIVE, PERCENTAGE_MEASURE, calculate_percentage
 from akvo.utils import ensure_decimal
