+++ akvo/rest/views/result.py
@@ -5,7 +5,7 @@ See more details in the license.txt file located at the root folder of the Akvo 
 For additional details on the GNU license please see < http://www.gnu.org/licenses/agpl.html >.
 """
 
-from akvo.rest.models import TastyTokenAuthentication, JWTAuthentication
+from akvo.rest.authentication import TastyTokenAuthentication, JWTAuthentication
 from akvo.rsr.models import Indicator, Result, Project
 from django.shortcuts import get_object_or_404
 from rest_framework.authentication import SessionAuthentication
